+++ src/std.rs
@@ -1,0 +1,380 @@
+pub use std::*;
+
+#[cfg(feature = "sanitize")]
+pub mod sync {
+    pub use std::sync::*;
+
+    /// A mutual exclusion primitive useful for protecting shared data
+    ///
+    /// This mutex will block threads waiting for the lock to become available. The
+    /// mutex can be created via a [`new`] constructor. Each mutex has a type parameter
+    /// which represents the data that it is protecting. The data can only be accessed
+    /// through the RAII guards returned from [`lock`] and [`try_lock`], which
+    /// guarantees that the data is only ever accessed when the mutex is locked.
+    ///
+    /// # Poisoning
+    ///
+    /// The mutexes in this module implement a strategy called "poisoning" where a
+    /// mutex is considered poisoned whenever a thread panics while holding the
+    /// mutex. Once a mutex is poisoned, all other threads are unable to access the
+    /// data by default as it is likely tainted (some invariant is not being
+    /// upheld).
+    ///
+    /// For a mutex, this means that the [`lock`] and [`try_lock`] methods return a
+    /// [`Result`] which indicates whether a mutex has been poisoned or not. Most
+    /// usage of a mutex will simply [`unwrap()`] these results, propagating panics
+    /// among threads to ensure that a possibly invalid invariant is not witnessed.
+    ///
+    /// A poisoned mutex, however, does not prevent all access to the underlying
+    /// data. The [`PoisonError`] type has an [`into_inner`] method which will return
+    /// the guard that would have otherwise been returned on a successful lock. This
+    /// allows access to the data, despite the lock being poisoned.
+    ///
+    /// [`new`]: Self::new
+    /// [`lock`]: Self::lock
+    /// [`try_lock`]: Self::try_lock
+    /// [`unwrap()`]: Result::unwrap
+    /// [`PoisonError`]: std::sync::PoisonError
+    /// [`into_inner`]: std::sync::PoisonError::into_inner
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::sync::{Arc, Mutex};
+    /// use std::thread;
+    /// use std::sync::mpsc::channel;
+    ///
+    /// const N: usize = 10;
+    ///
+    /// // Spawn a few threads to increment a shared variable (non-atomically), and
+    /// // let the main thread know once all increments are done.
+    /// //
+    /// // Here we're using an Arc to share memory among threads, and the data inside
+    /// // the Arc is protected with a mutex.
+    /// let data = Arc::new(Mutex::new(0));
+    ///
+    /// let (tx, rx) = channel();
+    /// for _ in 0..N {
+    ///     let (data, tx) = (Arc::clone(&data), tx.clone());
+    ///     thread::spawn(move || {
+    ///         // The shared state can only be accessed once the lock is held.
+    ///         // Our non-atomic increment is safe because we're the only thread
+    ///         // which can access the shared state when the lock is held.
+    ///         //
+    ///         // We unwrap() the return value to assert that we are not expecting
+    ///         // threads to ever fail while holding the lock.
+    ///         let mut data = data.lock().unwrap();
+    ///         *data += 1;
+    ///         if *data == N {
+    ///             tx.send(()).unwrap();
+    ///         }
+    ///         // the lock is unlocked here when `data` goes out of scope.
+    ///     });
+    /// }
+    ///
+    /// rx.recv().unwrap();
+    /// ```
+    ///
+    /// To recover from a poisoned mutex:
+    ///
+    /// ```
+    /// use std::sync::{Arc, Mutex};
+    /// use std::thread;
+    ///
+    /// let lock = Arc::new(Mutex::new(0_u32));
+    /// let lock2 = Arc::clone(&lock);
+    ///
+    /// let _ = thread::spawn(move || -> () {
+    ///     // This thread will acquire the mutex first, unwrapping the result of
+    ///     // `lock` because the lock has not been poisoned.
+    ///     let _guard = lock2.lock().unwrap();
+    ///
+    ///     // This panic while holding the lock (`_guard` is in scope) will poison
+    ///     // the mutex.
+    ///     panic!();
+    /// }).join();
+    ///
+    /// // The lock is poisoned by this point, but the returned result can be
+    /// // pattern matched on to return the underlying guard on both branches.
+    /// let mut guard = match lock.lock() {
+    ///     Ok(guard) => guard,
+    ///     Err(poisoned) => poisoned.into_inner(),
+    /// };
+    ///
+    /// *guard += 1;
+    /// ```
+    ///
+    /// To unlock a mutex guard sooner than the end of the enclosing scope,
+    /// either create an inner scope or drop the guard manually.
+    ///
+    /// ```
+    /// use std::sync::{Arc, Mutex};
+    /// use std::thread;
+    ///
+    /// const N: usize = 3;
+    ///
+    /// let data_mutex = Arc::new(Mutex::new(vec![1, 2, 3, 4]));
+    /// let res_mutex = Arc::new(Mutex::new(0));
+    ///
+    /// let mut threads = Vec::with_capacity(N);
+    /// (0..N).for_each(|_| {
+    ///     let data_mutex_clone = Arc::clone(&data_mutex);
+    ///     let res_mutex_clone = Arc::clone(&res_mutex);
+    ///
+    ///     threads.push(thread::spawn(move || {
+    ///         // Here we use a block to limit the lifetime of the lock guard.
+    ///         let result = {
+    ///             let mut data = data_mutex_clone.lock().unwrap();
+    ///             // This is the result of some important and long-ish work.
+    ///             let result = data.iter().fold(0, |acc, x| acc + x * 2);
+    ///             data.push(result);
+    ///             result
+    ///             // The mutex guard gets dropped here, together with any other values
+    ///             // created in the critical section.
+    ///         };
+    ///         // The guard created here is a temporary dropped at the end of the statement, i.e.
+    ///         // the lock would not remain being held even if the thread did some additional work.
+    ///         *res_mutex_clone.lock().unwrap() += result;
+    ///     }));
+    /// });
+    ///
+    /// let mut data = data_mutex.lock().unwrap();
+    /// // This is the result of some important and long-ish work.
+    /// let result = data.iter().fold(0, |acc, x| acc + x * 2);
+    /// data.push(result);
+    /// // We drop the `data` explicitly because it's not necessary anymore and the
+    /// // thread still has work to do. This allows other threads to start working on
+    /// // the data immediately, without waiting for the rest of the unrelated work
+    /// // to be done here.
+    /// //
+    /// // It's even more important here than in the threads because we `.join` the
+    /// // threads after that. If we had not dropped the mutex guard, a thread could
+    /// // be waiting forever for it, causing a deadlock.
+    /// // As in the threads, a block could have been used instead of calling the
+    /// // `drop` function.
+    /// drop(data);
+    /// // Here the mutex guard is not assigned to a variable and so, even if the
+    /// // scope does not end after this line, the mutex is still released: there is
+    /// // no deadlock.
+    /// *res_mutex.lock().unwrap() += result;
+    ///
+    /// threads.into_iter().for_each(|thread| {
+    ///     thread
+    ///         .join()
+    ///         .expect("The thread creating or execution failed !")
+    /// });
+    ///
+    /// assert_eq!(*res_mutex.lock().unwrap(), 800);
+    /// ```
+    ///
+    pub struct Mutex<T: ?Sized> {
+        inner: std::sync::Mutex<T>,
+    }
+
+    unsafe impl<T: ?Sized + Send> Send for Mutex<T> {}
+    unsafe impl<T: ?Sized + Send> Sync for Mutex<T> {}
+
+    impl<T> Mutex<T> {
+        /// Creates a new mutex in an unlocked state ready for use.
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use std::sync::Mutex;
+        ///
+        /// let mutex = Mutex::new(0);
+        /// ```
+        pub fn new(t: T) -> Self {
+            Self {
+                inner: std::sync::Mutex::new(t),
+            }
+        }
+
+        /// Acquires a mutex, blocking the current thread until it is able to do so.
+        ///
+        /// This function will block the local thread until it is available to acquire
+        /// the mutex. Upon returning, the thread is the only thread with the lock
+        /// held. An RAII guard is returned to allow scoped unlock of the lock. When
+        /// the guard goes out of scope, the mutex will be unlocked.
+        ///
+        /// The exact behavior on locking a mutex in the thread which already holds
+        /// the lock is left unspecified. However, this function will not return on
+        /// the second call (it might panic or deadlock, for example).
+        ///
+        /// # Errors
+        ///
+        /// If another user of this mutex panicked while holding the mutex, then
+        /// this call will return an error once the mutex is acquired.
+        ///
+        /// # Panics
+        ///
+        /// This function might panic when called if the lock is already held by
+        /// the current thread.
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use std::sync::{Arc, Mutex};
+        /// use std::thread;
+        ///
+        /// let mutex = Arc::new(Mutex::new(0));
+        /// let c_mutex = Arc::clone(&mutex);
+        ///
+        /// thread::spawn(move || {
+        ///     *c_mutex.lock().unwrap() = 10;
+        /// }).join().expect("thread::spawn failed");
+        /// assert_eq!(*mutex.lock().unwrap(), 10);
+        /// ```
+        pub fn lock(&self) -> LockResult<MutexGuard<T>> {
+            crate::notify_blocking_call("lock\0");
+            self.inner.lock()
+        }
+
+        /// Attempts to acquire this lock.
+        ///
+        /// If the lock could not be acquired at this time, then [`Err`] is returned.
+        /// Otherwise, an RAII guard is returned. The lock will be unlocked when the
+        /// guard is dropped.
+        ///
+        /// This function does not block.
+        ///
+        /// # Errors
+        ///
+        /// If another user of this mutex panicked while holding the mutex, then
+        /// this call will return the [`Poisoned`] error if the mutex would
+        /// otherwise be acquired.
+        ///
+        /// If the mutex could not be acquired because it is already locked, then
+        /// this call will return the [`WouldBlock`] error.
+        ///
+        /// [`Poisoned`]: TryLockError::Poisoned
+        /// [`WouldBlock`]: TryLockError::WouldBlock
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use std::sync::{Arc, Mutex};
+        /// use std::thread;
+        ///
+        /// let mutex = Arc::new(Mutex::new(0));
+        /// let c_mutex = Arc::clone(&mutex);
+        ///
+        /// thread::spawn(move || {
+        ///     let mut lock = c_mutex.try_lock();
+        ///     if let Ok(ref mut mutex) = lock {
+        ///         **mutex = 10;
+        ///     } else {
+        ///         println!("try_lock failed");
+        ///     }
+        /// }).join().expect("thread::spawn failed");
+        /// assert_eq!(*mutex.lock().unwrap(), 10);
+        /// ```
+        pub fn try_lock(&self) -> TryLockResult<MutexGuard<T>> {
+            self.inner.try_lock()
+        }
+
+        /// Determines whether the mutex is poisoned.
+        ///
+        /// If another thread is active, the mutex can still become poisoned at any
+        /// time. You should not trust a `false` value for program correctness
+        /// without additional synchronization.
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use std::sync::{Arc, Mutex};
+        /// use std::thread;
+        ///
+        /// let mutex = Arc::new(Mutex::new(0));
+        /// let c_mutex = Arc::clone(&mutex);
+        ///
+        /// let _ = thread::spawn(move || {
+        ///     let _lock = c_mutex.lock().unwrap();
+        ///     panic!(); // the mutex gets poisoned
+        /// }).join();
+        /// assert_eq!(mutex.is_poisoned(), true);
+        /// ```
+        #[inline]
+        pub fn is_poisoned(&self) -> bool {
+            self.inner.is_poisoned()
+        }
+
+        /// Clear the poisoned state from a mutex.
+        ///
+        /// If the mutex is poisoned, it will remain poisoned until this function is called. This
+        /// allows recovering from a poisoned state and marking that it has recovered. For example, if
+        /// the value is overwritten by a known-good value, then the mutex can be marked as
+        /// un-poisoned. Or possibly, the value could be inspected to determine if it is in a
+        /// consistent state, and if so the poison is removed.
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use std::sync::{Arc, Mutex};
+        /// use std::thread;
+        ///
+        /// let mutex = Arc::new(Mutex::new(0));
+        /// let c_mutex = Arc::clone(&mutex);
+        ///
+        /// let _ = thread::spawn(move || {
+        ///     let _lock = c_mutex.lock().unwrap();
+        ///     panic!(); // the mutex gets poisoned
+        /// }).join();
+        ///
+        /// assert_eq!(mutex.is_poisoned(), true);
+        /// let x = mutex.lock().unwrap_or_else(|mut e| {
+        ///     **e.get_mut() = 1;
+        ///     mutex.clear_poison();
+        ///     e.into_inner()
+        /// });
+        /// assert_eq!(mutex.is_poisoned(), false);
+        /// assert_eq!(*x, 1);
+        /// ```
+        #[inline]
+        pub fn clear_poison(&self) {
+            self.inner.clear_poison();
+        }
+
+        /// Consumes this mutex, returning the underlying data.
+        ///
+        /// # Errors
+        ///
+        /// If another user of this mutex panicked while holding the mutex, then
+        /// this call will return an error instead.
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use std::sync::Mutex;
+        ///
+        /// let mutex = Mutex::new(0);
+        /// assert_eq!(mutex.into_inner().unwrap(), 0);
+        /// ```
+        pub fn into_inner(self) -> LockResult<T> {
+            self.inner.into_inner()
+        }
+
+        /// Returns a mutable reference to the underlying data.
+        ///
+        /// Since this call borrows the `Mutex` mutably, no actual locking needs to
+        /// take place -- the mutable borrow statically guarantees no locks exist.
+        ///
+        /// # Errors
+        ///
+        /// If another user of this mutex panicked while holding the mutex, then
+        /// this call will return an error instead.
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use std::sync::Mutex;
+        ///
+        /// let mut mutex = Mutex::new(0);
+        /// *mutex.get_mut().unwrap() = 10;
+        /// assert_eq!(*mutex.lock().unwrap(), 10);
+        /// ```
+        pub fn get_mut(&mut self) -> Result<&mut T, std::sync::PoisonError<&mut T>> {
+            self.inner.get_mut()
+        }
+    }
+}
